refactor: Favorite 도메인 관련 피드백 반영(로그인 멤버의 유효성 검사)
